Adds a conditional for checking if click events have been defined yet on the virtual keyboard, even if the keyboard exists in the DOM
https://trello.com/c/JFg6Nbpk/87-9-24-21-dry-run-of-app-build-process

After downloading the site from Webflow, the virtual keyboard is actually compiled already in the HTML.   Whereas, on the web page, it is dynamically generated.

Therefore, in the Electron context, the virtual keyboard exists already in the DOM, and that causes the keyboard click events to not be set up.

I copied the source code from the virtual keyboard and then altered to make sure the click events are added, once and only once, even if the keyboard already exists in the DOM.

https://github.com/furcan/KioskBoard/blob/main/src/all-in-one/kioskboard-aio.js